Chunker outputs an iterator
Changed chunker to produce an iterator over Slices, with start and end
position context. This allows a file to be processed as it is read,
opening up the door to future optimizations like producing the next
slice in a future, enabling chunking to happen alongside the data
being processed by the consumer.

Implements #1