Implement manifest for FileSystem
Proved a less than ideal, nonconforming implementation of the maifest trait for the FileSystem
backend. It is currently very agressive with locking, overwrites the entire file every time a change
is made, and does not support multiple connections.

Added unit tests for FileSystem implementation of manifest trait.

Added method to StoredArchive to produce a dummy archive in the test build configuration.