Add new backend methods for async IO
Added the Backend::read_chunk and Backend::write_chunk methods from #29.
The backends have been updated to have the methods with an
unimplemented!() for the time being.

I decided to remove length from segment descriptor, and instead decided to
enforce that backends use a format that does not require prior knowlege
of the lengthof the chunk.