Applied empty-vec optimization to Transaction struct.
Having to encode the option in the transaction struct was resulting in a massive encoding overhead.
Since, in actual use, the contents will always be an encoded Chunk, which can not be zero length due
to the tags, this should not result in any acutal issues.

In the future, this should probably be swapped out to have the interface deal directly with chunks
instead of byte slices, with the option restored, to provide more consistentcy and avoid the current
sutiation of double encoding.