Removed async_std
I don't have anything against async_std, but the interplay between it and futures-rs was causing
substantial slowdown and, occasionally, livelocks. In addition, it is not really compatible with the
wasm-wasi target at this time.

This required making BackupDriver an async trait, which required minor api changes, mostly involving
passing owned values, as well as setting a bunch of trait bounds.

Tests have been updated to hit the new api as well.