Removed old prototype filesystem backend
Removed the old filesystem backend, as it is soon to be replaced with multifile, and its existance
is interfering with code clean up of superseeded traits and concepts elsewhere in the codebase.

Modified tests, where it makes sense to, to use the mem backend, and commenet out some tests that
will be restored when the multifile backend is implemented.