Implement Index for multifile format
Created a new index module in common for the new IndexTransaction type.

Removed the todo!() skeleton in the multifile module.

Created a new index module in multifile.

Implemeted a somewhat spec compliant index impl that wraps the currently being written index file in
its own task, and communicates with that task using a mpsc channel, implementing a lock-free
archetecture.

See: #21