Implement InternalSegmentHandler
Implemented the underlying, non-thread-safe segment handler that will be used for the multifile
backend. This will require chaging the mutability in the type signatures of several methods in the
backend trait, and propagating those outwards, so that a tasked implementation can be easily
implemented with MPMC channels, without relying on locks.