Migrate to tokio executor
Migrated form the futures::excutor::ThreadPool executor to the tokio runtime.

I don't like making this decsion, since it is not as generic, but considering the future of the
project, which will include lots of networked I/O, as well as the slight ergonomics gain of using
the tokio runtime, I consider this the best way foward.

Lots of tests have been reworked to be #[tokio::test]