Add and test LZMA compression
Added support for LZMA compression, and updated the chunk test to check it as well.

Effectivly closes #7. While I would like to add support for more compression types in the future,
the combo of LZ4, ZSTD, and LZMA already cover 99% of use cases, so this is more than good enough
for an initial release.