Implement fastcdc with new chunker API
Reimplemented the fastcdc chunker (with tests) in the new chunker crate, hitting the new chunking
API. This implemenation performs less copying, directly reading into the main buffer

Additionally provided an error implementation for the module via thiserror.

FastCDC could still use a specialized version of of the iterator for when it is fed an AsRef<[u8]>
rather than a reader.