fix(asuran): Fix a cache invalidation bug
The multifile segment handler was reusing existing segment handlers after they had been written to,
invalidating their internal cache.

Setup logic to evict segments from the read only cache if they are selected to be written to.
Add logic to flush the current writing segment if it is selected for reading.

Closes #54