fix(asuran): Make IV regeneration resistant to mishandling
Previously, the IV associated with an Encryption needed to be regenerated before using the value to
preform an action. This makes the Encryption API fragile and prone to dangerous misuse.

Change the encrypt and encrypt_bytes methods of Encryption to take a mutable reference, and
automatically generate a new IV before encryption, to make the API more resistance to misuse. This
makes it impossible to decide beforehand what IV will be used to encrypt a piece of data, but if you
are doing this, you are probably doing something wrong.

encryption and encrypt_bytes now change the provided Encryption to have the IV that was generated
for the operation.

Closes #65

BREAKING CHANGE: Encryption's encrypt and encrypt_bytes now require a mutable reference.